Slf4jFactory as toplevel trait
